Backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
python-jose
passlib
bcrypt
numpy
pandas
scikit-learn
statsmodels
prophet
tensorflow
joblib
google-generativeai
google-cloud-aiplatform
redis
ortools
python-multipart
ipykernel
seaborn
xgboost 
lightgbm
cartopy